requirements.txt
Click==7.0
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
itsdangerous==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.1
PyYAML==3.13
smmap2==2.0.5
Werkzeug==0.14.1
